v2.4
Now you can change date with keyboard

+++ Clipboard-Save-2_0/cbilistview.cpp
@@ -29,20 +29,24 @@ void CBIListView::slotMenuTriggered(QAction* ac) {
     QString i = ac->text(); // info of clicked action
     if(currentIndex().isValid()){
         if(i == "Info..."){
-            QDialog dio(NULL, Qt::WindowCloseButtonHint);
+            QDialog *dio = new QDialog(NULL, Qt::WindowCloseButtonHint);
+            CBI tempCBI = CBIModel->getCBI(currentIndex());
+            QTextEdit* txt = new QTextEdit;
+            QLabel *lbl = new QLabel(tempCBI.Time().toString("hh:mm"));
 
-            QTextEdit txt;
-            QLabel lbl(CBIModel->getCBI(currentIndex()).Time().toString("hh:mm"));
+            txt->setReadOnly(true);
+            txt->setText(tempCBI.Data());
 
-            txt.setReadOnly(true);
-            txt.setText(CBIModel->getCBI(currentIndex()).Data());
+            QVBoxLayout *vbl = new QVBoxLayout;
+            vbl->addWidget(txt);
+            vbl->addWidget(lbl);
+            dio->setLayout(vbl);
+            dio->exec();
 
-            QVBoxLayout vbl;
-            vbl.addWidget(&txt);
-            vbl.addWidget(&lbl);
-            dio.setLayout(&vbl);
-            dio.exec();
-            dio.done(0);
+            delete txt;
+            delete lbl;
+            delete vbl;
+            delete dio;
         } else if (i == "Copy") {
             qApp->clipboard()->setText(CBIModel->getCBI(currentIndex()).Data());
         } else if (i == "Delete") {
@@ -54,26 +58,20 @@ void CBIListView::slotMenuTriggered(QAction* ac) {
 
 // Events
 void CBIListView::keyPressEvent(QKeyEvent *ev) {
-    if(ev->key() == Qt::Key_Delete && currentIndex().isValid()){
-        CBIModel->deleteCBI(currentIndex().row());
-    } else if( ev->key() == Qt::Key_Return &&  currentIndex().isValid())  {
-        qApp->clipboard()->setText(CBIModel->getCBI(currentIndex()).Data());
-    } else if(ev->key() == Qt::Key_Up && currentIndex().isValid() && currentIndex().row() != 0) {
-        setCurrentIndex(CBIModel->index(currentIndex().row() - 1, 0));
-    } else if(ev->key() == Qt::Key_Down && currentIndex().isValid() && currentIndex().row() < CBIModel->Size()){
-        setCurrentIndex(CBIModel->index(currentIndex().row() + 1, 0));
-    } else if(ev->key() == Qt::Key_Up && !currentIndex().isValid() && !CBIModel->isEmpty()) {
-        setCurrentIndex(CBIModel->index(CBIModel->Size() - 1, 0));
-    } else if(ev->key() == Qt::Key_Escape) {
-        parentWidget()->hide();
-    } else if(ev->key() == Qt::Key_Escape){
-        static_cast<CBS>(parentWidget()).hide();
+    int key = ev->key();
+           if(key == Qt::Key_Delete && currentIndex().isValid()                                            ){ CBIModel->deleteCBI(currentIndex().row());
+    } else if(key == Qt::Key_Return && currentIndex().isValid()                                            ){ qApp->clipboard()->setText(CBIModel->getCBI(currentIndex()).Data());
+    } else if(key == Qt::Key_Up     && currentIndex().isValid() &&  currentIndex().row() != 0              ){ setCurrentIndex(CBIModel->index(currentIndex().row() - 1, 0));
+    } else if(key == Qt::Key_Down   && currentIndex().isValid() &&  currentIndex().row() < CBIModel->Size()){ setCurrentIndex(CBIModel->index(currentIndex().row() + 1, 0));
+    } else if(key == Qt::Key_Up     && !currentIndex().isValid() && !CBIModel->isEmpty()                   ){ setCurrentIndex(CBIModel->index(CBIModel->Size() - 1, 0));
+    } else if(key == Qt::Key_Escape                                                                        ){ parentWidget()->hide();
+    } else if(key == Qt::Key_Right                                                                         ){ static_cast<CBS*>(parentWidget())->NextBackDay(1);
+    } else if(key == Qt::Key_Left                                                                          ){ static_cast<CBS*>(parentWidget())->NextBackDay(0);
+    } else if(key == Qt::Key_C || key == 1057                                                              ){ static_cast<CBS*>(parentWidget())->setCurrentDate();
     }
 }
 
-void CBIListView::contextMenuEvent(QContextMenuEvent* ev) {
-    m->exec(ev->globalPos());
-}
+void CBIListView::contextMenuEvent(QContextMenuEvent* ev) { m->exec(ev->globalPos()); }
 
 void CBIListView::mouseDoubleClickEvent(QMouseEvent *ev) {
     Q_UNUSED(ev)
